Finalizes mat stacking
Mats now stack on eachother as long as they are fully on the parent mat,
with a 75 px margin. For an unknown reason, the right margin will not
work with the same code setup as the other three sides. Implementing the
same solution for the right side produces display error.